Revert "Improving the Custom Methods and Add Batch on get Method"
